+++ ml_tools/schema.py
@@ -1,8 +1,10 @@
 from ._core._schema import (
     FeatureSchema,
+    create_guischema_template,
     info
 )
 
 __all__ = [
-    "FeatureSchema"
+    "FeatureSchema",
+    "create_guischema_template",
 ]